moving exception class to data module
