+++ core/data/src/main/java/org/mozilla/social/core/data/FailedUpdateException.kt
@@ -1,3 +1,3 @@
-package org.mozilla.social.common
+package org.mozilla.social.core.data
 
 class FailedUpdateException(e: Exception) : Exception(e)
